Fix wrong channel type in Client/Server CommandFactories This is a u8 not a u16

+++ src/network/serveropcodes.h
@@ -41,7 +41,7 @@ struct ToServerCommandHandler
 struct ClientCommandFactory
 {
 	const char* name;
-	u16 channel;
+	u8 channel;
 	bool reliable;
 };
